+++ utils/tests/buffertest.cc
@@ -21,64 +21,64 @@ int main()
     
     printf("InPlaceBuffer TEST\n");
     printf("A %20u used, %u free, offset %d/%d\n",
-	   b.used(), b.free(), b.strchr('c'), b.strchr((unsigned char)'c'));
+	   (unsigned)b.used(), (unsigned)b.free(), (int)b.strchr('c'), (int)b.strchr((unsigned char)'c'));
 
     b.put("frogs on ice", 13);
     printf("B %20u used, %u free, offset %d/%d\n",
-	   b.used(), b.free(), b.strchr('c'), b.strchr((unsigned char)'c'));
+	   (unsigned)b.used(), (unsigned)b.free(), (int)b.strchr('c'), (int)b.strchr((unsigned char)'c'));
     
     s = (char *)b.get(8);
     printf("C s: %s\n", s);
     
     printf("D %20u used, %u free, offset %d/%d\n",
-	   b.used(), b.free(), b.strchr('c'), b.strchr((unsigned char)'c'));
+	   (unsigned)b.used(), (unsigned)b.free(), (int)b.strchr('c'), (int)b.strchr((unsigned char)'c'));
 
     s = (char *)b.get(5);
     printf("E s: %s\n", s);
     
     printf("F %20u used, %u free, offset %d/%d\n",
-	   b.used(), b.free(), b.strchr('c'), b.strchr((unsigned char)'c'));
+	   (unsigned)b.used(), (unsigned)b.free(), (int)b.strchr('c'), (int)b.strchr((unsigned char)'c'));
            
     b.zap();
     printf("G %20u used, %u free, offset %d/%d\n",
-	   b.used(), b.free(), b.strchr('c'), b.strchr((unsigned char)'c'));
+	   (unsigned)b.used(), (unsigned)b.free(), (int)b.strchr('c'), (int)b.strchr((unsigned char)'c'));
     printf("\n");
     
     printf("BUFFER TEST\n");
     
     printf("A %20u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
     
     bb.put("frogs on ice", 13);
     printf("B %20u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
     
     bb.put("frogs on rice", 14);
     printf("C %20u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
     
     s = (char *)bb.get(8);
     printf("D s: %s\n", s);  // "frogs on ice"
     
     printf("E %20u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
     
     bb.put("frogs on bryce", 15);
     s = (char *)bb.get(5);
     printf("F s: %s\n", s);  // " ice"
 
     printf("G %20u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
     
     s = (char *)bb.get(16);
     printf("H s: %s\n", s);  // "frogs on rice"
     
     printf("I %20u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
 
     bb.unget(12);
     printf("I2 %19u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
     
     s = (char *)bb.get(11);
     printf("J s: %s\n", s);  // "s on rice"
@@ -87,7 +87,7 @@ int main()
     printf("J2 s: %s\n", s);  // "rogs on bryce"
     
     printf("K %20u used, offset %d/%d\n",
-	   bb.used(), bb.strchr('c'), bb.strchr((unsigned char)'c'));
+	   (unsigned)bb.used(), (int)bb.strchr('c'), (int)bb.strchr((unsigned char)'c'));
     printf("\n");
     
     printf("BUFFER STRESS\n");
@@ -98,7 +98,7 @@ int main()
 	s = (char *)bb.alloc(i);
 	memcpy(s, xx, i);
 #ifdef DEBUG_STRESS
-        fprintf(stderr, "alloc(%d)\n", i);
+        fprintf(stderr, "alloc(%d)\n", (int)i);
 #endif
 	in += i;
 	total += i;
@@ -108,7 +108,7 @@ int main()
 	if (i > lastalloc)
 	    i = lastalloc;
 #ifdef DEBUG_STRESS
-        fprintf(stderr, "unalloc(%d)\n", i);
+        fprintf(stderr, "unalloc(%d)\n", (int)i);
 #endif
 	bb.unalloc(i);
 	in -= i;
@@ -117,14 +117,14 @@ int main()
 	if (i > in)
 	    i = in;
 #ifdef DEBUG_STRESS
-        fprintf(stderr, "get(%d)\n", i);
+        fprintf(stderr, "get(%d)\n", (int)i);
 #endif
 	bb.get(i);
 	in -= i;
 	
 	i = random() % sizeof(xx);
 #ifdef DEBUG_STRESS
-        fprintf(stderr, "put(%d)\n", i);
+        fprintf(stderr, "put(%d)\n", (int)i);
 #endif
 	bb.put(xx, i);
 	in += i;
@@ -134,7 +134,7 @@ int main()
 	if (i > in)
 	    i = in;
 #ifdef DEBUG_STRESS
-        fprintf(stderr, "get(%d)\n", i);
+        fprintf(stderr, "get(%d)\n", (int)i);
 #endif
 	bb.get(i);
 	in -= i;
@@ -144,7 +144,7 @@ int main()
 	if (i > lastput)
 	    i = lastput;
 #ifdef DEBUG_STRESS
-        fprintf(stderr, "unget(%d)\n", i);
+        fprintf(stderr, "unget(%d)\n", (int)i);
 #endif
 	bb.unget(i);
 	in += i;
@@ -154,10 +154,10 @@ int main()
 	{
 	    max = bb.used();
 	    printf("New max: %u bytes in subbuffers after %u bytes\n",
-		   max, total);
+		   (unsigned)max, (unsigned)total);
 	}
 #ifdef DEBUG_STRESS
-	fprintf(stderr, "[%6d]", in);
+	fprintf(stderr, "[%6d]", (int)in);
 #endif
     }
